Add problem 1015
And several syntax improvements

+++ problemas/1000/1001.php
@@ -1,6 +1,6 @@
 <?php
 
-$a = (int)readline();
-$b = (int)readline();
+$a = readline();
+$b = readline();
 $x = $a + $b;
 echo("X = {$x}\n");
