+++ problemas/1000/1002.php
@@ -1,6 +1,6 @@
 <?php
 
 $n = 3.14159;
-$raio = (float)readline();
+$raio = readline();
 $area = $n * $raio * $raio;
 echo("A=" . number_format($area, 4, '.', '') . "\n");
