+++ problemas/1000/1005.php
@@ -1,6 +1,6 @@
 <?php
 
-$a = (float)readline();
-$b = (float)readline();
+$a = readline();
+$b = readline();
 $media = ($a * 3.5 + $b * 7.5) / 11.0;
 echo("MEDIA = " . number_format($media, 5) . "\n");
