+++ problemas/1000/1006.php
@@ -1,7 +1,7 @@
 <?php
 
-$a = (float)readline();
-$b = (float)readline();
-$c = (float)readline();
+$a = readline();
+$b = readline();
+$c = readline();
 $media = ($a * 2 + $b * 3 + $c * 5) / 10.0;
 echo("MEDIA = " . number_format($media, 1) . "\n");
