+++ problemas/1000/1009.php
@@ -1,7 +1,7 @@
 <?php
 
 readline();
-$salario = (float)readline();
-$vendas = (float)readline();
+$salario = readline();
+$vendas = readline();
 $total = $salario + $vendas * 0.15;
 echo("TOTAL = R$ " . number_format($total, 2, '.', '') . "\n");
